telephony_api/pyproject.toml
[tool.poetry]
name = "telephony-api"
version = "0.1.0"
description = ""
authors = ["dudaborges <duda.pborges92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
tortoise-orm = "^0.21.7"
aerich = "^0.7.2"
asyncpg = "^0.30.0"
uvicorn = "^0.32.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
pre_test = "task lint"
test = "pytest -s -x --cov=telephony_api -vv"
post_test = "coverage html"